added dto's, added data validation for dtos, improved basic crud operations

+++ src/main/java/com/librarymanagement/LibraryManagement/author/AuthorService.java
@@ -1,6 +1,9 @@
 package com.librarymanagement.LibraryManagement.author;
 
 import com.librarymanagement.LibraryManagement.author.exception.AuthorNotFoundException;
+import com.librarymanagement.LibraryManagement.book.Book;
+import com.librarymanagement.LibraryManagement.book.BookRepository;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,33 +14,51 @@ import java.util.Set;
 
 @Service
 public class AuthorService {
+    private final BookRepository bookRepository;
     private final AuthorRepository authorRepository;
 
     @Autowired
-    public AuthorService(AuthorRepository authorRepository) {
+    public AuthorService(BookRepository bookRepository, AuthorRepository authorRepository) {
+        this.bookRepository = bookRepository;
         this.authorRepository = authorRepository;
     }
 
     // Create
     @Transactional
     public Author saveAuthor(Author author) {
-        Optional<Author> existingAuthor = authorRepository.findByFirstNameAndLastName(author.getFirstName(), author.getLastName());
+        // Szukanie autora po imieniu, nazwisku oraz innych atrybutach
+        Optional<Author> existingAuthor = authorRepository.findAuthorByFirstNameAndLastNameAndNationalityAndDateOfBirth(
+                author.getFirstName(),
+                author.getLastName(),
+                author.getNationality(),
+                author.getDateOfBirth()
+        );
 
         if (existingAuthor.isPresent()) {
             throw new IllegalArgumentException("Provided Author already exists");
         }
+
         return authorRepository.save(author);
     }
+
     // Read
     public List<Author> getAllAuthors() {
-        List<Author> authors = authorRepository.findAll();
 
-        return authors;
+        return authorRepository.findAll();
     }
 
     public Author getAuthorById(long id) {
         Optional<Author> author = authorRepository.findById(id);
         return author.orElseThrow(() -> new AuthorNotFoundException("Author with id " + id + " does not exist"));
+    }
+
+    public Optional<Author> getAuthorByFirstNameAndLastNameAndNationalityAndDateOfBirth(Author author) {
+        return authorRepository.findAuthorByFirstNameAndLastNameAndNationalityAndDateOfBirth(
+                author.getFirstName(),
+                author.getLastName(),
+                author.getNationality(),
+                author.getDateOfBirth()
+        );
     }
 
     // Update
@@ -76,8 +97,16 @@ public class AuthorService {
     // Delete
     @Transactional
     public void deleteAuthorById(long id) {
-        Author existingAuthor = findAuthorOrThrow(id);
-        authorRepository.delete(existingAuthor);
+        Optional<Author> optionalAuthor = authorRepository.findById(id);
+
+        if (optionalAuthor.isPresent()) {
+            Author existingAuthor = optionalAuthor.get();
+            bookRepository.deleteBooksByAuthorId(id);
+            authorRepository.delete(existingAuthor);  // Usuń autora
+            authorRepository.flush();  // Wymuś zapisanie zmian w bazie danych
+        } else {
+            throw new EntityNotFoundException("Author with id " + id + " not found");
+        }
     }
 
     // helper methods
